+++ src/gamspy/_container.py
@@ -649,6 +649,28 @@ class Container(gt.Container):
         self._options.miro_protect = original_state
         self._temp_container.data = {}
 
+    def _load_records_with_rename(
+        self, load_from: str, names: dict[str, str]
+    ) -> None:
+        self._temp_container.read(load_from, list(names.keys()))
+        original_state = self._options.miro_protect
+        self._options.miro_protect = False
+
+        for gdx_name, gamspy_name in names.items():
+            if gamspy_name in self.data:
+                updated_records = self._temp_container[gdx_name].records
+                self[gamspy_name].records = updated_records
+                self[gamspy_name].domain_labels = self[
+                    gamspy_name
+                ].domain_names
+            else:
+                raise ValidationError(
+                    f"Invalid renaming. `{gamspy_name}` does not exist in the container."
+                )
+
+        self._options.miro_protect = original_state
+        self._temp_container.data = {}
+
     def _read(
         self,
         load_from: str | Container | gt.Container,
@@ -766,7 +788,7 @@ class Container(gt.Container):
     def loadRecordsFromGdx(
         self,
         load_from: str,
-        symbol_names: Iterable[str] | None = None,
+        symbol_names: Iterable[str] | dict[str, str] | None = None,
     ) -> None:
         """
         Loads data of the given symbols from a GDX file. If no
@@ -776,8 +798,11 @@ class Container(gt.Container):
         ----------
         load_from : str
             Path to the GDX file
-        symbol_names : Iterable[str], optional
-            Symbols whose data will be load from GDX, by default None
+        symbol_names : Iterable[str], dict[str, str], optional
+            Symbol names whose data will be load from GDX, by default None.
+            Default option loads records of all symbols in the GDX file.
+            If given as a dict, keys are the symbol names in the GDX file, and
+            values are the names of the GAMSPy symbols.
 
         Examples
         --------
@@ -797,7 +822,11 @@ class Container(gt.Container):
                 self.system_directory, load_from
             )
 
-        self._load_records_from_gdx(load_from, symbol_names)
+        if isinstance(symbol_names, dict):
+            self._load_records_with_rename(load_from, symbol_names)
+        else:
+            self._load_records_from_gdx(load_from, symbol_names)
+
         self._synch_with_gams()
 
     def addGamsCode(self, gams_code: str) -> None: