+++ CHANGELOG.md
@@ -4,6 +4,7 @@ GAMSPy CHANGELOG
 GAMSPy 1.10.0
 -------------
 - General
+  - Allow renaming on container.loadRecordsFromGdx function call.
   - Allow users to disable all validation via gp.set_options({"VALIDATION": 0}).
 
 GAMSPy 1.9.0